Moving boss to be spawned by the enemy factory, and moved the tests to enemyfactory test as well
