+++ source/core/src/main/com/csse3200/game/areas/ForestGameArea.java
@@ -232,7 +232,7 @@ public class ForestGameArea extends GameArea {
 
     private void spawnKangarooBoss() {
         if (!kangarooBossSpawned) {
-            Entity kangarooBoss = NPCFactory.createKangaBossEntity(player);
+            Entity kangarooBoss = EnemyFactory.createKangaBossEntity(player);
             spawnEntityOnMap(kangarooBoss);
             kangarooBossSpawned = true;
         }
