+++ source/core/src/main/com/csse3200/game/entities/configs/PlayerConfig.java
@@ -3,7 +3,4 @@ package com.csse3200.game.entities.configs;
 /**
  * Defines the properties stored in player config files to be loaded by the Player Factory.
  */
-public class PlayerConfig extends BaseFriendlyEntityConfig  {
-  public int gold = 1;
-  public String favouriteColour = "none";
-}
+public class PlayerConfig extends BaseFriendlyEntityConfig  {}
