+++ source/core/src/test/com/csse3200/game/entities/factories/NPCFactoryTest.java
@@ -84,7 +84,6 @@ class NPCFactoryTest {
         eagle = NPCFactory.createEagle(player, enemies);
         turtle = NPCFactory.createTurtle(player, enemies);
         snake = NPCFactory.createSnake(player, enemies);
-        kanga = NPCFactory.createKangaBossEntity(player);
     }
 
     /**
@@ -608,76 +607,5 @@ class NPCFactoryTest {
         snake.setPosition(pos);
 
         assertEquals(pos, snake.getPosition());
-    }
-
-    static class TestComponent1 extends Component {}
-
-    /**
-     * Tests Creation of Kanga.
-     */
-    @Test
-    void TestKangaCreation() {
-        assertNotNull(kanga, "Kanga should not be null.");
-    }
-
-    /**
-     * Tests that the Kanga is an Entity.
-     */
-    @Test
-    void TestKangaIsEntity() {
-        assertEquals(kanga.getClass(), Entity.class);
-    }
-
-    /**
-     * Tests that the Kanga has a physics component.
-     */
-    @Test
-    void TestKangaHasPhysicsComponent() {
-        assertNotNull(kanga.getComponent(PhysicsComponent.class));
-    }
-
-    /**
-     * Tests that the Kanga has a physics movement component.
-     */
-    @Test
-    void TestKangaHasPhysicsMovementComponent() {
-        assertNotNull(kanga.getComponent(PhysicsMovementComponent.class));
-    }
-
-    /**
-     * Tests the Kanga has a collider component.
-     */
-    @Test
-    void TestKangaHasColliderComponent() {
-        assertNotNull(kanga.getComponent(ColliderComponent.class));
-    }
-
-    /**
-     * Tests that the Kanga has the correct HP stat.
-     */
-    @Test
-    void TestKangaHasCorrectHP() {
-        assertEquals(100, kanga.getComponent(CombatStatsComponent.class).getHealth(),
-                "Kanga should have 100 HP.");
-    }
-
-    /**
-     * Tests that the Kanga has an idle animation.
-     */
-    @Test
-    void TestKangaHasAnimation() {
-        assertTrue(kanga.getComponent(AnimationRenderComponent.class).hasAnimation("float") ,
-                "Kanga should have idle animation.");
-    }
-
-    /**
-     * Tests that the Kanga is in the correct spot when placed.
-     */
-    @Test
-    void TestKangaSetPosition() {
-        Vector2 pos = new Vector2(0f, 0f);
-        kanga.setPosition(pos);
-
-        assertEquals(pos, kanga.getPosition());
     }
 }
